+++ sdk/data/azcosmos/CHANGELOG.md
@@ -4,11 +4,16 @@
 
 ### Features Added
 
+* Adds support for float 16 datatype for vector embedding policy. See [PR 25707](https://github.com/Azure/azure-sdk-for-go/pull/25707)
+* Improved the performance of the built-in ReadMany implementation (when no QueryEngine is present). See [PR 26007](https://github.com/Azure/azure-sdk-for-go/pull/26007)
+
 ### Breaking Changes
 
 ### Bugs Fixed
 
 ### Other Changes
+
+* Small performance optimizations to API's using query engine. See [PR 25669](https://github.com/Azure/azure-sdk-for-go/pull/25669)
 
 ## 1.5.0-beta.4 (2025-11-24)